(lll￢ω￢)修复JsonObject输出string类型值的错误

+++ EntryEngine/Serialize/Json.cs
@@ -1301,7 +1301,7 @@ namespace EntryEngine.Serialize
                         case '\n': builder.Append("\\n"); break;
                         case '\r': builder.Append("\\r"); break;
                         case '\t': builder.Append("\\t"); break;
-                        default: builder.Append((ushort)c); break;
+                        default: builder.Append(c); break;
                     }
                 }
                 builder.Append('\"');
